Fixed a precision arithmetic bug which stopped the home position to be properly set at the start of the scene.
The issue being the initial home longitude and latitude were precise to 6 decimals but following queries of longitude and latitude, due to GPS noise were ~14 decimals precise or more. This caused the initial coordinates to appear much farther from the current longitude and latitude.

calling `NavigationUpdate` prior to calling SetHome corrects this